requirements.txt
cryptography
opencv-python
face_recognition
numpy
liboqs-python
python-pptx
